Move Noop Trace implementation to one class.
